simply move PicardSolve out of Executioner #9038
